--- packages/ui/src/GenericInputForm/FunctionParamLabel.tsx
@@ -1,20 +1,0 @@
-import { AbiParameter } from "viem";
-import { FC } from "react";
-import { Box, Text } from "@mantine/core";
-
-interface FunctionParamLabelProps {
-    input: AbiParameter;
-}
-
-export const FunctionParamLabel: FC<FunctionParamLabelProps> = ({ input }) => {
-    return (
-        <Box display="inline">
-            <Text c="cyan" span fz="sm">
-                {input.type}
-            </Text>{" "}
-            <Text span fz="sm">
-                {input.name}
-            </Text>
-        </Box>
-    );
-};
