+++ packages/ui/src/GenericInputForm/FunctionSignature.tsx
@@ -1,7 +1,23 @@
-import { AbiFunction } from "viem";
+import { AbiFunction, AbiParameter } from "viem";
 import { FC, Fragment } from "react";
-import { Text } from "@mantine/core";
-import { FunctionParamLabel } from "./FunctionParamLabel";
+import { Box, Text } from "@mantine/core";
+
+interface FunctionParamLabelProps {
+    input: AbiParameter;
+}
+
+export const FunctionParamLabel: FC<FunctionParamLabelProps> = ({ input }) => {
+    return (
+        <Box display="inline">
+            <Text c="cyan" span fz="sm">
+                {input.type}
+            </Text>{" "}
+            <Text span fz="sm">
+                {input.name}
+            </Text>
+        </Box>
+    );
+};
 
 interface FunctionSignatureProps {
     abiFunction: AbiFunction;
