--- packages/ui/src/GenericInputForm/initialValues.ts
@@ -1,14 +1,0 @@
-import { FormValues } from "./types";
-
-export const initialValues: FormValues = {
-    mode: "hex",
-    application: "",
-    rawInput: "0x",
-    stringInput: "",
-    abiMethod: "existing",
-    specificationMode: "json_abi",
-    humanAbi: "",
-    specificationId: "",
-    abiFunctionName: "",
-    abiFunctionParams: [],
-};
